platforms: xilinx: Add support for mapping PM nodes into 64-bit addresses
struct pm_access addr member is of type uint32_t which only allows mapping
PM nodes into 32-bit addresses. However, there are devices like
xlnx,ai-engine which are mapped above 32-bit space (ai-engine is mapped at
0x20000000000). In order to support such mappings, modify the addr type
to be uint64_t.

Currently memory mapped definition for Versal AI engine node (MM_DEV_AIE)
is set to 0xf70a0000. This is actually an address of AIE NPI space which
according to device tree bindings documentation is a temporary solution.
Set this address to 0x20000000000 instead, at which AIE is mapped, so it
can be properly binded to VERSAL_PM_DEV_AIE node (PM device node can only
be mapped to a single device so we cannot support both AIE and AIE NPI).

+++ xen/arch/arm/include/asm/platforms/xilinx-eemi.h
@@ -147,7 +147,7 @@ enum pm_ret_status {
  */
 struct pm_access
 {
-    uint32_t addr;
+    uint64_t addr;
     bool hwdom_access;    /* HW domain gets access regardless. */
 };
